requirements.txt
# Core Web Framework
# Note: Using compatible versions for Python 3.13 and Pydantic v1
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Instagram Scraping
# Note: Using both yt-dlp and instaloader for hybrid approach
# yt-dlp: Fast, no login required, good for single media
# instaloader: Better carousel support, requires login
yt-dlp==2024.10.22
instaloader==4.10.3

# HTTP & Async Operations
requests==2.31.0
httpx==0.25.1
aiofiles==23.2.1
Pillow>=10.0.0  # For thumbnail image processing

# Data Validation & Configuration
# Note: Using pydantic v1 for better Python 3.13 compatibility
pydantic==1.10.13
python-dotenv==1.0.0

# Rate Limiting & Caching
slowapi==0.1.9

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
